# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BusinessLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
    type: object
...
  dto.CoinProgramParticipantReturn:
    properties:
      balance:
        type: integer
      card_color:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.CoinProgramWithRewardsReturn:
    properties:
      coin_program:
        $ref: '#/definitions/dto.CoinProgramParticipantReturn'
      rewards:
        items:
          $ref: '#/definitions/dto.RewardReturn'
        type: array
    type: object
  dto.HTTPStatus:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  dto.RewardBuyResponse:
    properties:
      code:
        type: string
    type: object
...
  dto.RewardReturn:
    properties:
      cost:
        type: integer
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
...
  dto.UserGetResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      statistics:
        $ref: '#/definitions/dto.UserStatisticsResponse'
    type: object
  dto.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
    type: object
...
  dto.UserStatisticsResponse:
    properties:
      coupons_bought:
        type: integer
      user_qr_scanned_count:
        type: integer
    type: object
host: prod-team-22-t62v97db.final.prodcontest.ru
info:
  contact: { }
  description: API of LoyalT - platform for creating and managing loyalty programs
  termsOfService: http://swagger.io/terms/
  title: LoyalT API
  version: "1.0"
paths:
...
  /business/auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate business and set secure cookie "business_auth_token"
        with JWT token
      parameters:
        - description: Business credentials
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dto.BusinessLoginRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      summary: Business login
      tags:
        - Auth
...
  /user/auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate user and set secure cookie "user_auth_token" with
        JWT token
      parameters:
        - description: User credentials
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dto.UserLoginRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      summary: User login
      tags:
        - Auth
...
  /user/coin_programs:
    get:
      consumes:
        - application/json
      description: Get user's coin programs. Requires business_auth_token cookie for
        authentication.
      parameters:
        - description: Limit
          in: query
          name: limit
          required: true
          type: integer
        - description: Offset
          in: query
          name: offset
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CoinProgramParticipantReturn'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - UserAuthCookie: [ ]
      summary: List user's coin programs
      tags:
        - User
  /user/coin_programs/{coin_program_id}/rewards/{reward_id}/buy:
    post:
      consumes:
        - application/json
      description: Buy reward by its id. Requires user_auth_token cookie for authentication.
      parameters:
        - description: Coin program ID (uuid)
          in: path
          name: coin_program_participant_id
          required: true
          type: string
        - description: Reward ID (uuid)
          in: path
          name: reward_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RewardBuyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - UserAuthCookie: [ ]
      summary: Buy reward by its id
      tags:
        - User
  /user/coin_programs/{coin_program_participant_id}:
    get:
      consumes:
        - application/json
      description: Get user's coin program by ID. Requires business_auth_token cookie
        for authentication.
      parameters:
        - description: Coin program ID (uuid)
          in: path
          name: coin_program_participant_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CoinProgramParticipantReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - UserAuthCookie: [ ]
      summary: Get user's coin program by id
      tags:
        - User
  /user/coin_programs/{coin_program_participant_id}/rewards:
    get:
      consumes:
        - application/json
      description: List rewards for user by coin program id. Requires business_auth_token
        cookie for authentication.
      parameters:
        - description: Coin program participant id
          in: path
          name: coin_program_participant_id
          required: true
          type: string
        - default: 10
          description: Number of items to return
          in: query
          minimum: 1
          name: limit
          type: integer
        - default: 0
          description: Number of items to skip
          in: query
          minimum: 0
          name: offset
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: List of rewards
          schema:
            items:
              $ref: '#/definitions/dto.RewardReturn'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - UserAuthCookie: [ ]
      summary: List rewards for user
      tags:
        - User
  /user/coin_programs/rewards/available:
    get:
      consumes:
        - application/json
      description: Get rewards available for user. Requires business_auth_token cookie
        for authentication.
      parameters:
        - description: Limit
          in: query
          name: limit
          required: true
          type: integer
        - description: Offset
          in: query
          name: offset
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CoinProgramWithRewardsReturn'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - UserAuthCookie: [ ]
      summary: Get rewards available for user
      tags:
        - User
  /user/me:
    get:
      consumes:
        - application/json
      description: Get user profile
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserGetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      summary: Get user profile
      tags:
        - Auth
...
schemes:
  - https
swagger: "2.0"